refactoring interfaces/implementations
